Removed unused geometry methods
