+++ reconstruction/fvt/src/main/java/org/jlab/rec/fmt/hit/Hit.java
@@ -3,7 +3,7 @@ package org.jlab.rec.fmt.hit;
 import org.jlab.geom.prim.Line3D;
 import org.jlab.geom.prim.Point3D;
 import org.jlab.rec.fmt.Constants;
-import org.jlab.rec.fmt.GeometryMethods;
+import org.jlab.rec.fmt.Geometry;
 
 /**
  *
@@ -31,7 +31,7 @@ public class Hit implements Comparable<Hit>{
 		double x1 = Constants.FVT_stripsX[layer-1][strip-1][1];
 		double y0 = Constants.FVT_stripsY[layer-1][strip-1][0];
 		double y1 = Constants.FVT_stripsY[layer-1][strip-1][1];
-		double Z  = GeometryMethods.getLayerZ(layer-1);
+		double Z  = Geometry.getLayerZ(layer-1);
 		Line3D seg = new Line3D();
 		seg.setOrigin(new Point3D(x0,y0,Z));
 		seg.setEnd(new Point3D(x1,y1,Z));
